+++ src/Axent.Abstractions/IHandlerPipe.cs
@@ -1,8 +1,6 @@
 namespace Axent.Abstractions;
 
-#pragma warning disable S2326
-// ReSharper disable once UnusedTypeParameter
-public interface IHandlerPipe<TRequest, TResponse> : IAxentPipe<TRequest, TResponse> where TRequest : notnull
-#pragma warning restore S2326
-{
-}
+/// <summary>
+/// Marker interface for source generator
+/// </summary>
+public interface IHandlerPipe;