+++ src/Axent.Abstractions/IPipeline.cs
@@ -1,0 +1,6 @@
+namespace Axent.Abstractions;
+
+/// <summary>
+/// Marker interface for source generator
+/// </summary>
+public interface IPipeline;