+++ src/Axent.Core/DependencyInjection/AxentBuilderExtensions.cs
@@ -9,8 +9,17 @@ namespace Axent.Core.DependencyInjection;
 
 public static class AxentBuilderExtensions
 {
-    public static AxentBuilder AddAxent(this IServiceCollection services, Action<AxentOptions>? configure = null)
+    /// <summary>
+    /// Add the services required by axent
+    /// </summary>
+    /// <param name="services">ServiceCollection</param>
+    /// <param name="configure">Action to configure the <see cref="AxentOptions"/></param>
+    /// <param name="assemblies">Assemblies which get scanned for source generated handlers</param>
+    /// <returns></returns>
+    public static AxentBuilder AddAxent(this IServiceCollection services, Action<AxentOptions>? configure = null, Assembly[]? assemblies = null)
     {
+        assemblies ??= AppDomain.CurrentDomain.GetAssemblies();
+
         var builder = new AxentBuilder(services);
 
         var options = new AxentOptions();
@@ -35,7 +44,9 @@ public static class AxentBuilderExtensions
             builder.AddPipe(typeof(RequestLoggerPipe<,>));
         }
 
-        AxentSenderRegistry.Apply(services);
+        builder.AddSourceGeneratedServices(assemblies);
+
+
 
         return builder;
     }
@@ -120,5 +131,34 @@ public static class AxentBuilderExtensions
     {
         builder.Services.AddSingleton<ITransactionScopeFactory, TransactionScopeFactory>();
         builder.Services.AddScoped(typeof(ITransactionPipe<,>), typeof(TransactionPipe<,>));
+    }
+
+    private static AxentBuilder AddSourceGeneratedServices(
+        this AxentBuilder builder,
+        Assembly[] assemblies)
+    {
+        var senderType = typeof(ISender);
+        var pipelineType = typeof(IPipeline);
+        var handlerPipeType = typeof(IHandlerPipe);
+
+        var allTypes = assemblies.SelectMany(a => a.GetTypes())
+            .ToList();
+        var sender = allTypes.First(t => senderType.IsAssignableFrom(t));
+        builder.Services.AddScoped(senderType, sender);
+
+        foreach (var type in allTypes.Where(t => t is { IsAbstract: false, IsInterface: false }))
+        {
+            if (pipelineType.IsAssignableFrom(type))
+            {
+                builder.Services.AddScoped(type);
+            }
+
+            if (handlerPipeType.IsAssignableFrom(type))
+            {
+                builder.Services.AddScoped(type);
+            }
+        }
+
+        return builder;
     }
 }