+++ src/Axent.Extensions.Authorization/AxentBuilderExtensions.cs
@@ -1,0 +1,19 @@
+using Axent.Core.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Axent.Extensions.Authorization;
+
+public static class AxentBuilderExtensions
+{
+    public static AxentBuilder AddAuthorization(this AxentBuilder builder)
+    {
+        builder.Services.AddAuthorization();
+        builder.Services.AddHttpContextAccessor();
+
+        builder.Services.AddScoped<IPrincipalAccessor, HttpContextPrincipalAccessor>();
+
+        builder.AddPipe(typeof(AuthorizationPipe<,>));
+
+        return builder;
+    }
+}